requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.4
openai==0.28.0
openpyxl==3.1.5
pandas==2.2.3
piexif==1.1.3
pillow==11.2.1
pillow_heif==0.22.0
propcache==0.3.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
requests==2.32.3
six==1.17.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.19.0
